% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{quickColSums}
\alias{quickColSums}
\title{Add columns of sparse matrix and return a vector containing the sum of all column-wise elements}
\usage{
quickColSums(mat)
}
\arguments{
\item{mat}{A sparse matrix}
}
\description{
Add columns of sparse matrix and return a vector containing the sum of all column-wise elements
}
