\name{quickmatsums-package}
\alias{quickmatsums-package}
\alias{quickmatsums}
\docType{package}
\title{
\packageTitle{quickmatsums}
}
\description{
\packageDescription{quickmatsums}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{quickmatsums}
\packageIndices{quickmatsums}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{quickmatsums}

Maintainer: \packageMaintainer{quickmatsums}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
