% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{quickRowSums}
\alias{quickRowSums}
\title{Add rows of sparse matrix and return a vector containing the sum of all row-wise elements}
\usage{
quickRowSums(mat)
}
\arguments{
\item{mat}{A sparse matrix}
}
\description{
Add rows of sparse matrix and return a vector containing the sum of all row-wise elements
}
